feat(main.js-styles.css-index.html-classes.ts-classes.js): added class based data handling for clearer data passing. added vehicle selection to DOM
although this does not really change much (I am just using the old code's objects to supply the arguments for the new class objects) this does make a better seperation of concerns as the html generation is now linked directly to the data source rather than in a different function in charge of generating all the html. the decision for ts was mostly to ensure the proper enforcement of inheritance and data types to make it at least work internally to the modules. since it seems that the generated js is almost identical to the ts, it will likely not be worth the time and effort to add the ts generation to the ci pipeline, although I will really miss seeing all the typedefs.